+++ src/main/java/net/yxiao233/ifeu/client/data/ModBlockStateProvider.java
@@ -32,6 +32,7 @@ public class ModBlockStateProvider extends BlockStateProvider {
         blockItem(ModBlocks.RULE_CONTROLLER.getLeft());
         blockItem(ModContents.ULTIMATE_MACHINE_FRAME);
         blockItem(ModContents.DRAGON_STAR_BLOCK);
+        blockItem(ModBlocks.FLUID_CRAFTING_TABLE.getLeft());
     }
 
     private void blockItem(RegistryObject<Block> registryObject){