+++ src/main/java/net/yxiao233/ifeu/client/events/BlockRendererEvent.java
@@ -1,0 +1,19 @@
+package net.yxiao233.ifeu.client.events;
+
+import net.minecraft.world.level.block.entity.BlockEntityType;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.client.event.EntityRenderersEvent;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.fml.common.Mod;
+import net.yxiao233.ifeu.IndustrialForegoingExtraUpgrades;
+import net.yxiao233.ifeu.common.block.entity.FluidCraftingTableEntity;
+import net.yxiao233.ifeu.common.block.renderer.FluidCraftingTableRenderer;
+import net.yxiao233.ifeu.common.registry.ModBlocks;
+
+@Mod.EventBusSubscriber(modid = IndustrialForegoingExtraUpgrades.MODID,bus = Mod.EventBusSubscriber.Bus.MOD,value = Dist.CLIENT)
+public class BlockRendererEvent {
+    @SubscribeEvent
+    public static void registerBlockEntityRenderer(EntityRenderersEvent.RegisterRenderers event){
+        event.registerBlockEntityRenderer((BlockEntityType<FluidCraftingTableEntity>) ModBlocks.FLUID_CRAFTING_TABLE.getRight().get(),FluidCraftingTableRenderer::new);
+    }
+}